% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_present}
\alias{assert_present}
\title{Assert that required values are present in a vector}
\usage{
assert_present(available, required)
}
\arguments{
\item{available}{Vector of available values}

\item{required}{Vector of required values that must be present}
}
\value{
NULL invisibly; stops with error if any required values are missing
}
\description{
Assert that required values are present in a vector
}
\keyword{internal}
