% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_to_vec.R
\name{clip_to_vec}
\alias{clip_to_vec}
\title{Translate copied excel row/column into code to make an R vector}
\usage{
clip_to_vec()
}
\value{
Nothing
}
\description{
Function interacts with the computer clipboard.
To use, start by copying a row or column in excel that you want coded as a
vector in R. Run clip_to_vec() to change the system clipboard to
the string of R code that defines an equivalent vector, then paste into
R script.
}
