% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_files}
\alias{copy_files}
\title{Copy files to a destination directory}
\usage{
copy_files(source_files, dest_dir, overwrite = FALSE, timestamp = FALSE)
}
\arguments{
\item{source_files}{data.frame with columns \code{file_path} and \code{file_name}}

\item{dest_dir}{Destination directory (created if it doesn't exist)}

\item{overwrite}{Logical, whether to overwrite existing files}

\item{timestamp}{Logical, whether to add timestamp prefix to copied files}
}
\value{
NULL invisibly
}
\description{
Copy files to a destination directory
}
\keyword{internal}
