% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelsior-package.R
\docType{package}
\name{excelsior-package}
\alias{excelsior}
\alias{excelsior-package}
\title{excelsior: Streamlines Excel File Automation}
\description{
Useful tools for automation of excel file data entry, especially transfering parts of one sheet into a sheet of another file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cbedwards-dfw/excelsior}
  \item \url{https://cbedwards-dfw.github.io/excelsior/}
  \item \url{https://framverse.r-universe.dev/excelsior,}
  \item Report bugs at \url{https://github.com/cbedwards-dfw/excelsior/issues}
}

}
\author{
\strong{Maintainer}: Collin Edwards \email{Collin.Edwards@dfw.wa.gov} (\href{https://orcid.org/0000-0002-4937-5159}{ORCID})

}
\keyword{internal}
