% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_handling.R
\name{find_fill_color}
\alias{find_fill_color}
\title{Find fill color of specified cell}
\usage{
find_fill_color(filepath, sheet, address)
}
\arguments{
\item{filepath}{Excel file}

\item{sheet}{Excel sheet}

\item{address}{Cell address}
}
\value{
Excel rgb color (character vector of length 1). To use in R, add "#" to beginning.
}
\description{
Find fill color of specified cell
}
\examples{
\dontrun{
find_fill_color(primary_file, "D-3", "A103")
}
}
