% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sheet_dimensions}
\alias{get_sheet_dimensions}
\title{Get dimensions of data in a worksheet}
\usage{
get_sheet_dimensions(wb, sheet)
}
\arguments{
\item{wb}{Workbook object}

\item{sheet}{Sheet name or index}
}
\value{
Named list with max_row and max_col
}
\description{
Reads the worksheet to determine actual data extent.
}
\keyword{internal}
