% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_file_open}
\alias{is_file_open}
\title{Check if a file is currently open by another process}
\usage{
is_file_open(file_path)
}
\arguments{
\item{file_path}{Path to file to check}
}
\value{
Logical, TRUE if file is open/locked
}
\description{
Check if a file is currently open by another process
}
\keyword{internal}
