% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_to_col}
\alias{num_to_col}
\title{Convert numeric index to Excel column letters}
\usage{
num_to_col(num)
}
\arguments{
\item{num}{Integer representing column number (1-16384)}
}
\value{
Character string representing Excel column
}
\description{
Convert numeric index to Excel column letters
}
\keyword{internal}
