% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel.R
\name{read_excel}
\alias{read_excel}
\title{Read Excel workbooks into R}
\usage{
read_excel(file_names_df, reqd_sheets)
}
\arguments{
\item{file_names_df}{data.frame containing columns \code{file_path} and \code{file_name}}

\item{reqd_sheets}{string vector with the names of the worksheets to read. Use NA for all sheets.}
}
\value{
tbl_df with columns \code{file_name} and \code{raw_excel_data} (a nested named list indexed by sheet name)
}
\description{
Reads multiple Excel workbooks and returns raw cell data in a tidy format.
}
