% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel.R
\name{read_excel_single}
\alias{read_excel_single}
\title{Read a single Excel workbook}
\usage{
read_excel_single(file_path, reqd_sheets)
}
\arguments{
\item{file_path}{string indicating the file path to the workbook}

\item{reqd_sheets}{the sheets to read from the workbook. Use NA to read all sheets.}
}
\value{
a named list containing raw Excel data in a tbl_df, indexed by sheet name
}
\description{
Reads raw cell data from an Excel workbook using tidyxl.
}
