% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-xldiff.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{cell_range_translate}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{xldiff}{\code{\link[xldiff]{cell_range_translate}}}
}}

