% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_handling.R
\name{row_finder_by_color}
\alias{row_finder_by_color}
\title{Find anchor cell based on fill color}
\usage{
row_finder_by_color(
  filepath,
  sheet,
  column = 1,
  color = "FF000000",
  instance = 1,
  offset = 0
)
}
\arguments{
\item{filepath}{Character. Filepath to Excel file of interest}

\item{sheet}{Character or numeric. Sheet name or index to search in}

\item{column}{Character or numeric. Column that contains anchor cell as number (e.g., 3) or corresponding excel column label ("C").}

\item{color}{Character vector. Color of anchor cell in hex, with or without preliminary "#". Defaults to black ("FF000000").}

\item{instance}{Numeric. Which match to use if multiple matches found? Defaults to 1.}

\item{offset}{Numeric. Number to add to returned row number. Useful when pattern identifies a title row and you want the row below (1) or above (-1). Defaults to 0}
}
\value{
Row number (numeric of length 1)
}
\description{
Useful for handling sheets that use color-filled rows as separators (e.g., STT table files).
}
\examples{
\dontrun{
## find the row before the first column has a cell with black fill
row_finder_by_color("Appendix D.xlsx", sheet = "D-3", column = 1, offset = -1)
}
}
