% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_workbook_path}
\alias{validate_workbook_path}
\title{Validate workbook path}
\usage{
validate_workbook_path(wb_dir, must_exist = FALSE)
}
\arguments{
\item{wb_dir}{Path to workbook file}

\item{must_exist}{Logical, whether file must already exist}
}
\value{
Normalized path (invisibly)
}
\description{
Validate workbook path
}
\keyword{internal}
