% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_columns.R
\name{copy_columns}
\alias{copy_columns}
\title{Copy columns from one Excel file to another with change tracking}
\usage{
copy_columns(
  wb,
  to_file,
  from_file,
  sheet,
  columns,
  row_start,
  row_end,
  color_scheme,
  meta_cell,
  verbose = TRUE
)
}
\arguments{
\item{wb}{An openxlsx2 workbook object to modify}

\item{to_file}{Character. Filepath for the "original" reference file}

\item{from_file}{Character. Filepath containing the workbook with data to copy FROM}

\item{sheet}{Character. Sheet name to copy from/to}

\item{columns}{Character vector. Excel column letters to copy (e.g., c("A", "B"))}

\item{row_start}{Integer. Initial row number to copy}

\item{row_end}{Integer. Final row number to copy}

\item{color_scheme}{Character vector of length 2. Hex colors for "Changed" and "Unchanged" cells. Use the same general color with different shades.}

\item{meta_cell}{Character. Excel cell reference to add transfer metadata (color, file, date).}

\item{verbose}{Logical. Whether to print progress messages. Defaults to TRUE.}
}
\value{
Modified workbook object with copied data, change highlighting, and metadata
}
\description{
Copies specified columns from an input Excel file to a workbook, highlighting
changes compared to a master file and adding metadata about the transfer. \code{row_start} and \code{row_end} are most easily identified using \code{\link[=row_finder]{row_finder()}}.
}
\examples{
\dontrun{
wb <- load_workbook("primary.xlsx)
wb <- copy_columns(wb,
                  to_file = "primary.xlsx",
                  from_file = "update.xlsx",
                  sheet = "Sheet1",
                  columns = c("A", "B"),
                  row_start = 2,
                  row_end = 100,
                  color_scheme = c("#82DE4E", "#D5F4C3"),
                  meta_cell =  "D1")
}
}
\seealso{
\code{\link[=row_finder]{row_finder()}}
}
