% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_excel_data_tbl.R
\name{get_excel_data_tbl}
\alias{get_excel_data_tbl}
\title{get single tibbles per file for all required data given lists of tbl_dfs extracted from excel}
\usage{
get_excel_data_tbl(raw_excel_data_tbl)
}
\arguments{
\item{raw_excel_data_tbl}{tbl with file_name and nested list of raw excel data}
}
\value{
tbl_df with col file_name, nested raw excel data containing named lists of tibbles per sheet, nested tibbles containing combined data
}
\description{
get single tibbles per file for all required data given lists of tbl_dfs extracted from excel
}
