% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_excel_data_tbl.R
\name{get_excel_data_tbl_single}
\alias{get_excel_data_tbl_single}
\title{convert list of tbl_dfs extracted from excel to a tibble containing all extracted data}
\usage{
get_excel_data_tbl_single(raw_excel_data)
}
\arguments{
\item{raw_excel_data}{\code{list} of \code{data.frame}'s containing raw excel data where each list object contains data from a sheet required from the workbook}
}
\value{
\code{tbl_df} with columns sheet, row, col, numeric, character and date
}
\description{
convert list of tbl_dfs extracted from excel to a tibble containing all extracted data
}
