% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel.R
\name{read_excel}
\alias{read_excel}
\title{read excel workbooks in to R}
\usage{
read_excel(file_names_df, reqd_sheets)
}
\arguments{
\item{file_names_df}{\code{data.frame} containing the directories of files you'd like to read}

\item{reqd_sheets}{string vector with the names of the worksheets to read in}
}
\value{
\code{tbl_df} with file_name and raw_excel_data as a nested named list where the list is indexed by the sheet name specified in reqd_sheets
}
\description{
read excel workbooks in to R
}
