% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel.R
\name{read_excel_single}
\alias{read_excel_single}
\title{read a single excel workbook}
\usage{
read_excel_single(file_path, reqd_sheets)
}
\arguments{
\item{file_path}{string indicating the file path to the workbook of interest}

\item{reqd_sheets}{the sheets you'd like to copy from the workbook}
}
\value{
a named list containing raw excel data in a tbl_df where the name equals the reqd_sheet
}
\description{
read a single excel workbook
}
