% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wb.R
\name{read_wb}
\alias{read_wb}
\title{read data from workbooks}
\usage{
read_wb(
  source_data_dir,
  reqd_wkbks = NA,
  reqd_sheets = NA,
  password_protected = FALSE,
  overwrite = TRUE,
  dest_data_dir = NULL
)
}
\arguments{
\item{source_data_dir}{string indicating source directory}

\item{reqd_wkbks}{character vector indicating the names of the files you wish to extract from source_dir. default NA extracts all workbooks in the dir}

\item{reqd_sheets}{character vector containing names of sheets to extract data from. default NA extracts from all sheets in workbook}

\item{password_protected}{logical indicating whether the excel files are password protected. Password must be the same for multiple file reads}

\item{overwrite}{logical indicating whether files in the destination directory should be overwritten. FALSE results in no copy rather than duplicates}

\item{dest_data_dir}{optional string indicating destination directory - default NULL will use temp directory}
}
\value{
\code{tbl_df} with columns file_name, raw_excel_data and all_excel_data_tbl
}
\description{
read data from workbooks
}
