% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlock_wb.R
\name{unlock_wb}
\alias{unlock_wb}
\title{unlock a protected excel workbook}
\usage{
unlock_wb(file_dir, options.java.params = NULL)
}
\arguments{
\item{file_dir}{\code{data.frame} of file paths and names for files you wish to unlock}

\item{options.java.params}{optional string argument specifying java parameters}
}
\description{
unlock a protected excel workbook
}
