% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_wb.R
\name{write_wb.default}
\alias{write_wb.default}
\title{Default method for writing to excel}
\usage{
\method{write_wb}{default}(
  r_data,
  sheet_name,
  paste_coord = "A1",
  clear_sheet = TRUE,
  wb = NULL,
  wb_dir = NULL,
  save_wb = FALSE,
  ...
)
}
\arguments{
\item{r_data}{an object to be written to an xlsx workbook when r_data is a list it will use the names of the list as sheet_names}

\item{sheet_name}{a string indicating the name of the sheet to write to}

\item{paste_coord}{string indicating the coordinates of the top left cell you wish to write data to - default "A1"}

\item{clear_sheet}{logical indicating whether you'd like to clear the contents of the destination sheet before writing - default TRUE}

\item{wb}{workbook object - default NULL}

\item{wb_dir}{string indicating the workbook's directory}

\item{save_wb}{logical indicating whether workbook should be saved}

\item{...}{arguments for other methods}
}
\description{
Default method for writing to excel
}
