% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_wb.R
\name{write_wb.list}
\alias{write_wb.list}
\title{Write an object of class \code{list} to excel}
\usage{
\method{write_wb}{list}(
  r_data,
  clear_sheet = TRUE,
  wb = NULL,
  wb_dir = NULL,
  save_wb = FALSE,
  ...
)
}
\arguments{
\item{r_data}{an object of class \code{list}}

\item{clear_sheet}{logical indicating whether you'd like to clear the contents of the destination sheet before writing - default TRUE}

\item{wb}{workbook object - default NULL}

\item{wb_dir}{string indicating the workbook's directory}

\item{save_wb}{logical indicating whether workbook should be saved}

\item{...}{arguments for other methods}
}
\description{
Write an object of class \code{list} to excel
}
