% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_wb.R
\name{write_wb}
\alias{write_wb}
\title{Write from R to Excel}
\usage{
write_wb(r_data, ...)
}
\arguments{
\item{r_data}{an object to be written to an xlsx workbook when r_data is a list it will use the names of the list as sheet_names}

\item{...}{arguments for other methods}
}
\description{
Write from R to Excel
}
