% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_wb.R
\name{write_wb.zoo}
\alias{write_wb.zoo}
\title{Write an object of class \code{zoo} or \code{xts} to excel}
\usage{
\method{write_wb}{zoo}(
  r_data,
  sheet_name,
  paste_coord = "A1",
  clear_sheet = TRUE,
  wb = NULL,
  wb_dir = NULL,
  save_wb = FALSE,
  ...
)
}
\arguments{
\item{r_data}{an object of class \code{zoo}}

\item{sheet_name}{a string indicating the name of the sheet to write to}

\item{paste_coord}{string indicating the coordinates of the top left cell you wish to write data to - default "A1"}

\item{clear_sheet}{logical indicating whether you'd like to clear the contents of the destination sheet before writing - default TRUE}

\item{wb}{workbook object - default NULL}

\item{wb_dir}{string indicating the workbook's directory}

\item{save_wb}{logical indicating whether workbook should be saved}

\item{...}{arguments for other methods}
}
\description{
Write an object of class \code{zoo} or \code{xts} to excel
}
