% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_wb.R
\name{write_wb_multi}
\alias{write_wb_multi}
\title{Write nested tibbles to multiple workbooks}
\usage{
write_wb_multi(write_tbl, write_dir, ...)
}
\arguments{
\item{write_tbl}{an object of class \code{tbl_df} with columns name, object and data}

\item{write_dir}{string indicating the directory to write data to}

\item{...}{additional arguments to other methods}
}
\description{
Write nested tibbles to multiple workbooks
}
